Fixed appendix a bit, added some of my table results.
This is for the deconvolution procedure. I think we should have more detailed tables, and then from looking at these systematically, we should find the good distribution to use and the target variance to use. I have an idea to have maybe six different sub-tables :) That's what the supplementary material is for