Whew, redid the conservative korattikara test
According to discussions with john and myself. Now let's run the conservative
korattiakra test. with error bound of 0.1 we get minibatch size of 450 and an
epsilon of 0.01

er ... actually, if that's the case, there's no way their method will be better
because they are using 450 points per iteration, which automatically makes ours
faster. Hmmm ... maybe I should try an error rate of 0.05 instead? That means
eps = 0.001 and mb size of 50. Hmmm ... not sure. Let's do 0.05 becuase
otherwise we're giving ourselves too much advantage.